requirements.txt
asgiref==3.6.0
cryptography==41.0.3
cffi==1.15.1
bson==0.5.10

Django==3.2.22
django-cors-headers==3.5.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.11.2
django-dotenv==1.4.2

rq==1.15.1
django-rq==2.8.1

redis==4.5.4
django-redis==5.2.0


# Dependency injection
dependency-injector==4.41.0


dnslib==0.9.19
dnspython==1.15.0
mysqlclient==2.1.0
pytz==2020.4
pyjwt==2.6.0
sqlparse==0.4.4
gunicorn==19.10.0
gspread==5.3.2
slack_sdk==3.0.0rc1
pyhumps==1.6.1

requests==2.31.0
stripe==2.40.0

user-agents==2.0

schedule==0.6.0
tldextract==3.1.0
firebase-admin==5.2.0
pika==1.3.0
boto3==1.24.35
XlsxWriter==3.0.3
beautifulsoup4==4.12.2
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
oauth2client<4.0.0

pyotp==2.2.7

# Channels
channels==3.0.5
channels_redis==3.4.1
daphne==3.0.2

# Mailer
sendgrid==6.9.7
Jinja2==3.1.2

# Image
Pillow==10.0.1

# GeoIP
geoip2==4.7.0

